spark-runtime: split the HSS disk-ref checks into their own file
Pure move, no oracle change: basic.rs was at 463 lines and the oracle
repairs that follow need more room than the workspace 500-LoC cap leaves.
The four Issue #17 disk-ref checks move to tests/basic_hss.rs, nested via
the same #[path] pattern the snapshot index tests use, so they now
register as radix_tree::tests::basic::hss::*.

(cherry picked from commit 831adf661d371627e97de07dac5ac8afb7c556a3)
